feat(New-inventory-Layout): Added a new Layout for inventory
