+++ app/inventory/orders/page.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const Order = () => {
+  return <div>Order</div>
+}
+
+export default Order
