--- constants/bunnycdn/constant.tsx
@@ -1,4 +1,0 @@
-export const USERNAME = 'storehub'
-export const HOSTNAME = 'storage.bunnycdn.com'
-export const CONNECTION_TYPE = 'Passive'
-export const PORT = 21
